+++ src/probe/mod.rs
@@ -1,9 +1,11 @@
 pub mod correlate;
 pub mod icmp;
 pub mod socket;
+pub mod tcp;
 pub mod udp;
 
 pub use correlate::*;
 pub use icmp::*;
 pub use socket::*;
+pub use tcp::*;
 pub use udp::*;